k-means_img.py: Change `self.clusters` data struct
Instead of assign the vector to each cluster list,
just store the image vector's index that in list_image.

Because we need the index to access the list_lable array to
compute the accuracy.

There are another advantag--without copy vector, can save memory.
(do not sure whether python will copy the vector, whatever)